Refactoring Issues and PR widget into a common class
