+++ src/GitHubExtension/Widgets/GitHubWidget.cs
@@ -48,33 +48,9 @@ public abstract class GitHubWidget : WidgetImpl
 
     protected string SavedConfigurationData { get; set; } = string.Empty;
 
-    protected string RepositoryUrl { get; set; } = string.Empty;
-
-    protected string DeveloperLoginId { get; set; } = string.Empty;
-
     protected DateTime LastUpdated { get; set; } = DateTime.MinValue;
 
     protected DataUpdater DataUpdater { get; set; }
-
-    public string GetOwner()
-    {
-        return Validation.ParseOwnerFromGitHubURL(RepositoryUrl);
-    }
-
-    public string GetRepo()
-    {
-        return Validation.ParseRepositoryFromGitHubURL(RepositoryUrl);
-    }
-
-    public string GetIssueQuery()
-    {
-        return Validation.ParseIssueQueryFromGitHubURL(RepositoryUrl);
-    }
-
-    public string GetUnescapedIssueQuery()
-    {
-        return Uri.UnescapeDataString(GetIssueQuery()).Replace('+', ' ');
-    }
 
     public GitHubWidget()
     {
@@ -157,10 +133,6 @@ public abstract class GitHubWidget : WidgetImpl
 
         switch (verb)
         {
-            case WidgetAction.CheckUrl:
-                HandleCheckUrl(actionInvokedArgs);
-                break;
-
             case WidgetAction.SignIn:
                 _ = HandleSignIn();
                 break;
@@ -204,44 +176,12 @@ public abstract class GitHubWidget : WidgetImpl
         {
             return;
         }
-
-        RepositoryUrl = dataObject["url"]?.GetValue<string>() ?? string.Empty;
-        DeveloperLoginId = dataObject["account"]?.GetValue<string>() ?? string.Empty;
     }
 
     public override void OnCustomizationRequested(WidgetCustomizationRequestedArgs customizationRequestedArgs)
     {
         SavedConfigurationData = ConfigurationData;
         SetConfigure();
-    }
-
-    private void HandleCheckUrl(WidgetActionInvokedArgs args)
-    {
-        // Set loading page while we fetch data from GitHub.
-        Page = WidgetPageState.Loading;
-        UpdateWidget();
-
-        // This is the action when the user clicks the submit button after entering a URL while in
-        // the Configure state.
-        Page = WidgetPageState.Configure;
-        var data = args.Data;
-        var dataObject = JsonObject.Parse(data);
-        if (dataObject != null && dataObject["url"] != null)
-        {
-            RepositoryUrl = dataObject["url"]?.GetValue<string>() ?? string.Empty;
-            DeveloperLoginId = dataObject["account"]?.GetValue<string>() ?? string.Empty;
-
-            ConfigurationData = data;
-
-            var updateRequestOptions = new WidgetUpdateRequestOptions(Id)
-            {
-                Data = GetConfiguration(RepositoryUrl),
-                CustomState = ConfigurationData,
-                Template = GetTemplateForPage(Page),
-            };
-
-            WidgetManager.GetDefault().UpdateWidget(updateRequestOptions);
-        }
     }
 
     private async Task HandleSignIn()
@@ -253,7 +193,7 @@ public abstract class GitHubWidget : WidgetImpl
         Log.Logger()?.ReportInfo($"User sign in successful from WidgetAction invocation");
     }
 
-    private WidgetAction GetWidgetActionForVerb(string verb)
+    protected WidgetAction GetWidgetActionForVerb(string verb)
     {
         try
         {
@@ -264,103 +204,6 @@ public abstract class GitHubWidget : WidgetImpl
             // Invalid verb.
             Log.Logger()?.ReportError($"Unknown WidgetAction verb: {verb}");
             return WidgetAction.Unknown;
-        }
-    }
-
-    protected void AddDevIds(ref JsonObject configurationData)
-    {
-        var developerIdsData = new JsonArray();
-
-        foreach (var developerId in DeveloperIdProvider.GetInstance().GetLoggedInDeveloperIds().DeveloperIds)
-        {
-            Log.Logger()?.ReportInfo(developerId.LoginId);
-            developerIdsData.Add(new JsonObject
-            {
-                { "devId", developerId.LoginId },
-            });
-        }
-
-        configurationData.Add("accounts", developerIdsData);
-    }
-
-    public string GetConfiguration(string dataUrl)
-    {
-        var configurationData = new JsonObject
-        {
-            { "submitIcon", IconLoader.GetIconAsBase64("arrow.png") },
-        };
-
-        AddDevIds(ref configurationData);
-
-        if (dataUrl == string.Empty)
-        {
-            CanPin = false;
-            configurationData.Add("hasConfiguration", false);
-            configurationData.Add("configuring", !CanPin);
-            var repositoryData = new JsonObject
-            {
-                { "url", string.Empty },
-            };
-
-            configurationData.Add("configuration", repositoryData);
-            configurationData.Add("savedRepositoryUrl", SavedConfigurationData);
-            configurationData.Add("saveEnabled", false);
-
-            return configurationData.ToString();
-        }
-        else
-        {
-            try
-            {
-                // Get client for logged in user.
-                var client = GitHubClientProvider.Instance.GetClientForLoggedInDeveloper(true).Result;
-                if (client == null)
-                {
-                    throw new InvalidOperationException("Failed getting GitHubClient.");
-                }
-
-                // Get repository for the URL, which is "data" in this case.
-                var ownerName = Validation.ParseOwnerFromGitHubURL(dataUrl);
-                var repositoryName = Validation.ParseRepositoryFromGitHubURL(dataUrl);
-                var repository = client.Repository.Get(ownerName, repositoryName).Result;
-
-                var repositoryData = new JsonObject
-                {
-                    { "name", repository.FullName },
-                    { "label", repository.Name },
-                    { "owner", repository.Owner.Login },
-                    { "milestone", string.Empty },
-                    { "project", repository.Description },
-                    { "url", repository.HtmlUrl },
-                    { "query", GetUnescapedIssueQuery() },
-                };
-
-                configurationData.Add("hasConfiguration", true);
-                configurationData.Add("configuration", repositoryData);
-                configurationData.Add("savedRepositoryUrl", SavedConfigurationData);
-                configurationData.Add("saveEnabled", true);
-                CanPin = true;
-            }
-            catch (Exception ex)
-            {
-                Log.Logger()?.ReportError(Name, ShortId, $"Failed getting configuration information for input url: {dataUrl}", ex);
-                CanPin = false;
-                configurationData.Add("hasConfiguration", false);
-                configurationData.Add("configuring", !CanPin);
-
-                var repositoryData = new JsonObject
-                {
-                    { "url", RepositoryUrl },
-                };
-
-                configurationData.Add("errorMessage", ex.Message);
-                configurationData.Add("configuration", repositoryData);
-                configurationData.Add("saveEnabled", false);
-
-                return configurationData.ToString();
-            }
-
-            return configurationData.ToJsonString();
         }
     }
 
@@ -379,19 +222,6 @@ public abstract class GitHubWidget : WidgetImpl
     {
         IDeveloperIdProvider authProvider = DeveloperIdProvider.GetInstance();
         return authProvider.GetLoggedInDeveloperIds().DeveloperIds.Any();
-    }
-
-    protected IDeveloperId? GetWidgetDeveloperId()
-    {
-        foreach (var devid in DeveloperIdProvider.GetInstance().GetLoggedInDeveloperIds().DeveloperIds)
-        {
-            if (devid.LoginId == DeveloperLoginId)
-            {
-                return devid;
-            }
-        }
-
-        return null;
     }
 
     public void UpdateActivityState()
@@ -468,7 +298,7 @@ public abstract class GitHubWidget : WidgetImpl
 
     protected string GetCurrentState()
     {
-        return $"State: {ActivityState}  Page: {Page}  Data: {DataState}  Repository: {RepositoryUrl}";
+        return $"State: {ActivityState}  Page: {Page}  Data: {DataState}";
     }
 
     protected void LogCurrentState()
